requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
iniconfig==1.1.1
MouseInfo==0.1.3
numpy==1.20.2
opencv-python==4.5.1.48
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pytesseract==0.3.7
pytest==6.2.3
pytest-html==3.1.1
python-dotenv==0.17.0
PyTweening==1.0.3
scipy==1.6.3
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.4
